added pages folder and moved all page-components into that folder, updated relative path
