+++ client/pages/home-page.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import IconCards from './icon-cards';
-import SentenceStrip from './sentence-strip';
-import BottomNavBar from './bottom-nav-bar';
+import IconCards from '../components/icon-cards';
+import SentenceStrip from '../components/sentence-strip';
+import BottomNavBar from '../components/bottom-nav-bar';
 
 export default class HomePage extends React.Component {
   constructor(props) {
